+++ services/worker/src/worker/job_runners/dataset/split_names.py
@@ -5,14 +5,13 @@ import logging
 from http import HTTPStatus
 
 from libcommon.constants import CONFIG_SPLIT_NAMES_KINDS
+from libcommon.dtos import FullConfigItem, FullSplitItem
 from libcommon.exceptions import PreviousStepFormatError
 from libcommon.simple_cache import get_best_response, get_previous_step_or_raise
 
 from worker.dtos import (
     DatasetSplitNamesResponse,
     FailedConfigItem,
-    FullConfigItem,
-    FullSplitItem,
     JobResult,
 )
 from worker.job_runners.dataset.dataset_job_runner import DatasetJobRunner
