+++ services/worker/src/worker/job_runners/split/image_url_columns.py
@@ -3,15 +3,12 @@
 
 import logging
 
+from libcommon.dtos import SplitFirstRowsResponse
 from libcommon.exceptions import PreviousStepFormatError
 from libcommon.simple_cache import get_previous_step_or_raise
 from libcommon.utils import is_image_url
 
-from worker.dtos import (
-    CompleteJobResult,
-    ImageUrlColumnsResponse,
-    SplitFirstRowsResponse,
-)
+from worker.dtos import CompleteJobResult, ImageUrlColumnsResponse
 from worker.job_runners.split.split_job_runner import SplitJobRunner
 
 STRING_FEATURE_DTYPE = "string"
